requirements.txt
django==3.2
django-environ==0.4.5
psycopg2==2.8.6
gunicorn==20.1.0
djangorestframework==3.12.4
